fix(tests) Add pre-commit hook
Add a pre-commit hook and a pylint check.
This is to catch errors that were missed in testing before
you commit a change.

More checks will be added in the future.

Use 'cp tests/pre-commit .git/hooks' to activate.
Make sure that you have execute accss to the file, on Linux
that should already be the case.

tests/pre-commit: (New file)
  A pre-commit hook to be copied into .git/hooks

tests/pylint_check.sh:
  A pylint checking script